More imporvements to the model

+++ isteppable.h
@@ -2,7 +2,7 @@
 #define ISTEPPABLE_H_INCLUDED
 
 struct ISteppable {
-    virtual void step(int elapsed) = 0;
+    virtual void step(float elapsed) = 0;
 };
 
 #endif // ISTEPPABLE_H_INCLUDED
